move process_trace to sk-store
