+++ sk-skel/src/tests/ast_test.rs
@@ -1,6 +1,7 @@
 use std::collections::HashSet;
 
 use json_patch_ext::PointerBuf;
+use pest::Parser;
 use serde_json::{
     Value,
     json,
