+++ sk-skel/src/tests/mod.rs
@@ -1,6 +1,5 @@
 mod ast_test;
 mod engine_test;
-mod itest;
 mod parser_test;
 use assertables::*;
 use rstest::rstest;
