+++ sk-store/Cargo.toml
@@ -21,9 +21,11 @@ serde_json = { workspace = true }
 serde_yaml = { workspace = true }
 sk-api = { workspace = true }
 sk-core = { workspace = true }
+sk-skel = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
+pest = { workspace = true }
 
 [dev-dependencies]
 assertables = { workspace = true }
