+++ sk-store/src/tests/mod.rs
@@ -1,5 +1,6 @@
 mod exported_trace_test;
 mod import_export_test;
+mod itest;
 mod manager_test;
 mod pod_owners_map_test;
 mod trace_store_test;
